批量生成二维码时，统计成功和失败个数

+++ marketing-tool-service/src/main/java/cn/rongcapital/mkt/service/QrcodeCreateCountService.java
@@ -6,5 +6,5 @@ import cn.rongcapital.mkt.vo.BaseOutput;
  * Created by chengjincheng on 2016-8-15.
  */
 public interface QrcodeCreateCountService {
-	BaseOutput getCreateCount(Long batch_id);
+	BaseOutput getCreateCount(Integer batch_id);
 }
